rename src/filter.py into src/alpaka_filter.py
