+++ script/job_generator2/src/alpaka_bashi/__init__.py
@@ -10,7 +10,7 @@ from alpaka_bashi.versions import (
     get_alpaka_version_relation,
 )
 from alpaka_bashi.globals import *  # pylint: disable=wildcard-import,unused-wildcard-import
-from alpaka_bashi.filter import AlpakaFilter
+from alpaka_bashi.alpaka_filter import AlpakaFilter
 from alpaka_bashi.verify import verify
 from alpaka_bashi.runtime_info import get_runtime_infos
 from alpaka_bashi.combination import add_combinations_parameters
