+++ script/job_generator2/src/alpaka_bashi/alpaka_filter.py
