+++ script/job_generator2/tests/backend_filter/test_all_backends_fine.py
@@ -10,16 +10,16 @@ import unittest
 from collections import OrderedDict
 from bashi.globals import *  # pylint: disable=wildcard-import,unused-wildcard-import
 from utils import parse_param_value_tuples
-import alpaka_bashi.filter
+from alpaka_bashi.alpaka_filter import all_backends_fine
 
 
 class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
     def test_empty_row_empty_backend_list(self):
-        self.assertTrue(alpaka_bashi.filter.all_backends_fine(OrderedDict(), [], []))
+        self.assertTrue(all_backends_fine(OrderedDict(), [], []))
 
     def test_empty_row_single_backend(self):
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 row=OrderedDict(),
                 backends=[ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
                 all_available_backends=[ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
@@ -44,20 +44,16 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
     def test_empty_row_two_backend(self):
         for enabled_backends, all_backends in self.TEST_DATA_EMPTY_ROW_TWO_BACKEND:
             with self.subTest(enabled_backends=enabled_backends, all_backends=all_backends):
-                self.assertTrue(
-                    alpaka_bashi.filter.all_backends_fine(
-                        OrderedDict(), enabled_backends, all_backends
-                    )
-                )
+                self.assertTrue(all_backends_fine(OrderedDict(), enabled_backends, all_backends))
 
     def test_single_row_empty_backend_list(self):
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, OFF)]), [], []
             )
         )
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, OFF)]),
                 [],
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
@@ -66,7 +62,7 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
 
     def test_single_row_single_backend_valid(self):
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, ON)]),
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
@@ -74,7 +70,7 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
         )
 
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, ON)]),
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
                 [],
@@ -83,14 +79,14 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
 
     def test_single_row_single_backend_invalid(self):
         self.assertFalse(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, OFF)]),
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
             )
         )
         self.assertFalse(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, ON)]),
                 [],
                 [ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE],
@@ -99,7 +95,7 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
 
     def test_cuda_backend(self):
         self.assertFalse(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_GPU_CUDA_ENABLE, OFF)]),
                 [ALPAKA_ACC_GPU_CUDA_ENABLE],
                 [ALPAKA_ACC_GPU_CUDA_ENABLE],
@@ -107,7 +103,7 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
         )
 
         self.assertTrue(
-            alpaka_bashi.filter.all_backends_fine(
+            all_backends_fine(
                 parse_param_value_tuples([(ALPAKA_ACC_GPU_CUDA_ENABLE, 11.4)]),
                 [ALPAKA_ACC_GPU_CUDA_ENABLE],
                 [ALPAKA_ACC_GPU_CUDA_ENABLE],
@@ -199,7 +195,7 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
         for row, backends, all_backends in self.TEST_DATA_MULTIPLE_ROW_MULTIPLE_BACKEND_VALID_CASE:
             with self.subTest(row=row, backends=backends, all_backends=all_backends):
                 self.assertTrue(
-                    alpaka_bashi.filter.all_backends_fine(row, backends, all_backends),
+                    all_backends_fine(row, backends, all_backends),
                     f"\nrow:\n  {row}\nbackends:\n  {backends}\nall backends:  \n{all_backends}",
                 )
 
@@ -272,6 +268,6 @@ class TestAlpakaGetValidBackendCombinations(unittest.TestCase):
         ) in self.TEST_DATA_MULTIPLE_ROW_MULTIPLE_BACKEND_INVALID_CASE:
             with self.subTest(row=row, backends=backends, all_backends=all_backends):
                 self.assertFalse(
-                    alpaka_bashi.filter.all_backends_fine(row, backends, all_backends),
+                    all_backends_fine(row, backends, all_backends),
                     f"\nrow:\n  {row}\nbackends:\n  {backends}\nall backends:  \n{all_backends}",
                 )
