+++ script/job_generator2/tests/backend_filter/test_get_valid_backend_combinations.py
@@ -9,7 +9,7 @@ Custom filter for alpaka specific filter rules.
 import unittest
 from bashi.globals import *  # pylint: disable=wildcard-import,unused-wildcard-import
 from utils import parse_param_value_tuples
-import alpaka_bashi.filter
+from alpaka_bashi.alpaka_filter import get_valid_compiler_backend_combinations
 import bashi
 
 
@@ -17,7 +17,7 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
     def test_enabled_serial_cpu_is_always_required(self):
         input_row = parse_param_value_tuples([(ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, OFF)])
         self.assertEqual(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)),
+            len(get_valid_compiler_backend_combinations(input_row)),
             0,
             "Each compiler backend combination requires an enabled serial CPU backend. "
             "Therefore no combination with a disabled serial cpu backend can exist.",
@@ -27,33 +27,25 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
         input_row = parse_param_value_tuples(
             [(DEVICE_COMPILER, GCC, 8), (ALPAKA_ACC_CPU_B_SEQ_T_SEQ_ENABLE, ON)]
         )
-        self.assertGreater(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)), 0
-        )
+        self.assertGreater(len(get_valid_compiler_backend_combinations(input_row)), 0)
 
     def test_nvcc_compiler_and_hip_backend(self):
         input_row = parse_param_value_tuples(
             [(DEVICE_COMPILER, NVCC, 12.4), (ALPAKA_ACC_GPU_HIP_ENABLE, ON)]
         )
-        self.assertEqual(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)), 0
-        )
+        self.assertEqual(len(get_valid_compiler_backend_combinations(input_row)), 0)
 
     def test_gcc_host_and_icpx_backend(self):
         input_row = parse_param_value_tuples(
             [(HOST_COMPILER, GCC, 12), (ALPAKA_ACC_ONEAPI_GPU_ENABLE, ON)]
         )
-        self.assertEqual(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)), 0
-        )
+        self.assertEqual(len(get_valid_compiler_backend_combinations(input_row)), 0)
 
     def test_hipcc_compiler_and_disabled_cpu_backend(self):
         input_row = parse_param_value_tuples(
             [(DEVICE_COMPILER, HIPCC, 6.3), (ALPAKA_ACC_CPU_B_SEQ_T_THREADS_ENABLE, OFF)]
         )
-        self.assertGreater(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)), 0
-        )
+        self.assertGreater(len(get_valid_compiler_backend_combinations(input_row)), 0)
 
         input_row_2 = parse_param_value_tuples(
             [
@@ -62,9 +54,7 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
                 (ALPAKA_ACC_GPU_HIP_ENABLE, ON),
             ]
         )
-        self.assertGreater(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row_2)), 0
-        )
+        self.assertGreater(len(get_valid_compiler_backend_combinations(input_row_2)), 0)
 
     def test_icpx_compiler_and_different_sycl_backends_invalid(self):
         for sycl_backend in ONE_API_BACKENDS:
@@ -85,13 +75,13 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
                     + [(sycl_backend, ON)]
                 )
                 self.assertEqual(
-                    len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)),
+                    len(get_valid_compiler_backend_combinations(input_row)),
                     1,
                 )
         # icpx can be used for the three different one api backends
         self.assertEqual(
             len(
-                alpaka_bashi.filter.get_valid_compiler_backend_combinations(
+                get_valid_compiler_backend_combinations(
                     parse_param_value_tuples(
                         [(HOST_COMPILER, ICPX, "2025.0.4")],
                     )
@@ -109,7 +99,7 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
             ]
         )
         self.assertEqual(
-            len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(input_row)),
+            len(get_valid_compiler_backend_combinations(input_row)),
             2,  # host compiler GCC and Clang
         )
 
@@ -130,13 +120,13 @@ class TestAlpakaFilterValidGetBackendCombination(unittest.TestCase):
                 # The last element of parameter_values invalidates all possible backend combinations.
                 if len(row) < len(self.GROWING_ROW):
                     self.assertGreater(
-                        len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(row)),
+                        len(get_valid_compiler_backend_combinations(row)),
                         0,
                         f"\nrow: {bashi.get_str_row_nice(row)}",
                     )
                 else:
                     self.assertEqual(
-                        len(alpaka_bashi.filter.get_valid_compiler_backend_combinations(row)),
+                        len(get_valid_compiler_backend_combinations(row)),
                         0,
                         f"\nrow: {bashi.get_str_row_nice(row)}",
                     )
