+++ script/job_generator2/tests/test_alpaka_filter.py
@@ -13,7 +13,7 @@ from typing import Dict, Callable, cast
 import bashi
 from bashi.globals import *  # pylint: disable=wildcard-import,unused-wildcard-import
 from bashi.version.dependencies.clang_cuda import ClangCudaSDKSupport
-import alpaka_bashi.filter
+from alpaka_bashi.alpaka_filter import AlpakaFilter
 import alpaka_bashi.runtime_info
 from alpaka_bashi.globals import (
     RT_HOST_COMPILER_CUDA_SUPPORT,
@@ -50,9 +50,7 @@ class TestAlpakaFilter(unittest.TestCase):
     def test_valid_backend_combinations_a1(self):
         for row in self.VALID_BACKEND_COMBINATIONS:
             with self.subTest(row=row):
-                self.assertTrue(
-                    alpaka_bashi.filter.AlpakaFilter()(parse_param_value_tuples(row)), f"{row}"
-                )
+                self.assertTrue(AlpakaFilter()(parse_param_value_tuples(row)), f"{row}")
 
     INVALID_BACKEND_COMBINATIONS = [
         [(ALPAKA_ACC_GPU_HIP_ENABLE, ON), (ALPAKA_ACC_ONEAPI_GPU_ENABLE, ON)],
@@ -75,9 +73,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -121,9 +117,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg, runtime_infos=runtime_info)(
-                        row
-                    ),
+                    AlpakaFilter(output=reason_msg, runtime_infos=runtime_info)(row),
                 )
                 self.assertEqual(
                     reason_msg.getvalue(),
@@ -144,9 +138,7 @@ class TestAlpakaFilter(unittest.TestCase):
     def test_valid_hipcc62_debug_build_a3(self):
         for row in self.VALID_HIPCC_BUILD_CONFIGURATIONS:
             with self.subTest(row=row):
-                self.assertTrue(
-                    alpaka_bashi.filter.AlpakaFilter()(parse_param_value_tuples(row)), f"{row}"
-                )
+                self.assertTrue(AlpakaFilter()(parse_param_value_tuples(row)), f"{row}")
 
     INVALID_HIPCC_BUILD_CONFIGURATIONS = [
         [(DEVICE_COMPILER, HIPCC, 6.2), (BUILD_TYPE, CMAKE_DEBUG)],
@@ -158,9 +150,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -196,9 +186,7 @@ class TestAlpakaFilter(unittest.TestCase):
     def test_valid_ubuntu2204_a4(self):
         for row in self.VALID_AVAILABLE_COMPILERS_ON_UBUNTU:
             with self.subTest(row=row):
-                self.assertTrue(
-                    alpaka_bashi.filter.AlpakaFilter()(parse_param_value_tuples(row)), f"{row}"
-                )
+                self.assertTrue(AlpakaFilter()(parse_param_value_tuples(row)), f"{row}")
 
     INVALID_AVAILABLE_COMPILERS_ON_UBUNTU = [
         [(DEVICE_COMPILER, GCC, 14), (UBUNTU, "22.04")],
@@ -215,9 +203,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -235,9 +221,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -258,9 +242,7 @@ class TestAlpakaFilter(unittest.TestCase):
                 row = [(backend, ON), (UBUNTU, "22.04")]
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -277,9 +259,7 @@ class TestAlpakaFilter(unittest.TestCase):
     def test_valid_clang_ubuntu2204_a5(self):
         for row in self.AVAILABLE_CLANG_VERSIONS_ON_UBUNTU_2204:
             with self.subTest(row=row):
-                self.assertTrue(
-                    alpaka_bashi.filter.AlpakaFilter()(parse_param_value_tuples(row)), f"{row}"
-                )
+                self.assertTrue(AlpakaFilter()(parse_param_value_tuples(row)), f"{row}")
 
     def test_invalid_clang_ubuntu2204_wrong_ubuntu_a5(self):
         for clang_version, ubuntu_version in itertools.product((14, 16), ("24.4", "26.4")):
@@ -290,9 +270,7 @@ class TestAlpakaFilter(unittest.TestCase):
                 with self.subTest(row=row):
                     reason_msg = io.StringIO()
                     self.assertFalse(
-                        alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                            parse_param_value_tuples(row)
-                        ),
+                        AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                         f"{row}",
                     )
                     self.assertEqual(
@@ -311,9 +289,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -333,9 +309,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
                 self.assertEqual(
@@ -349,7 +323,7 @@ class TestAlpakaFilter(unittest.TestCase):
         row = (HOST_COMPILER, CLANG, 14), (ALPAKA_ACC_CPU_B_TBB_T_SEQ_ENABLE, OFF)
         reason_msg = io.StringIO()
         self.assertFalse(
-            alpaka_bashi.filter.AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
+            AlpakaFilter(output=reason_msg)(parse_param_value_tuples(row)),
             f"{row}",
         )
         self.assertEqual(
@@ -391,9 +365,7 @@ class TestAlpakaFilter(unittest.TestCase):
         for row in self.VALID_CLANG_CUDA_SDK_COMBINATIONS:
             with self.subTest(row=row):
                 self.assertTrue(
-                    alpaka_bashi.filter.AlpakaFilter(runtime_infos=runtime_info)(
-                        parse_param_value_tuples(row)
-                    ),
+                    AlpakaFilter(runtime_infos=runtime_info)(parse_param_value_tuples(row)),
                     f"{row}",
                 )
 
@@ -418,7 +390,7 @@ class TestAlpakaFilter(unittest.TestCase):
             with self.subTest(row=row):
                 reason_msg = io.StringIO()
                 self.assertFalse(
-                    alpaka_bashi.filter.AlpakaFilter(output=reason_msg, runtime_infos=runtime_info)(
+                    AlpakaFilter(output=reason_msg, runtime_infos=runtime_info)(
                         parse_param_value_tuples(row)
                     ),
                     f"{row}",
